requirements.txt
boto3==1.7.70
botocore==1.10.70
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
idna==2.7
jmespath==0.9.3
protobuf==3.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
